feat: implement dynamic social media previews for shared conversations
- Add server-side metadata generation with conversation-specific titles and descriptions
- Show first user message as title and AI responses with model names as description
- Support multi-AI response stitching (two models with 75 chars each)
- Add "AI Chat Sandbox" branding to all preview descriptions
- Maintain backward compatibility with existing conversation functionality
- Include comprehensive error handling and fallback metadata
- Add 1-hour caching for improved performance
